Get data from IDEEA database

Get data from IDEEA database

Usage

get_ideea_data(
  name,
  nreg = 7,
  variable,
  agg_fun = sum,
  raw = FALSE,
  offshore = FALSE,
  islands = FALSE
)

Arguments

name

character, name of data table, one of: "coal", "oil", "gas", "biomass", "lignite", ...

nreg

integer, number of region to return

variable

name or a part of name (regular expression) of variable to return.

agg_fun

character, function to aggregate data by region, default "sum" (for volumes), "mean" is advised for costs variable.

raw

logical, should the raw table be returned, FALSE by default

offshore

logical, should the data for offshore areas be returned, FALSE by default

islands

logical, should the data for remote islands be returned, FALSE by default

Examples

  get_ideea_data("coal", raw = T)
#> # A tibble: 34 × 13
#>    reg36 name36   mainland offshore reg1  name1 total_reserve_Mt production_2021
#>  * <chr> <chr>    <lgl>    <lgl>    <chr> <chr>            <dbl>           <dbl>
#>  1 AP    Andhra … TRUE     FALSE    IND   India           4142.           NA    
#>  2 AR    Arunach… TRUE     FALSE    IND   India             90.2          NA    
#>  3 AS    Assam    TRUE     FALSE    IND   India            525.            0.036
#>  4 BR    Bihar    TRUE     FALSE    IND   India           4437.           NA    
#>  5 CH    Chandig… TRUE     FALSE    IND   India             NA            NA    
#>  6 CT    Chhatti… TRUE     FALSE    IND   India          74192.          158.   
#>  7 DD    Daman a… TRUE     FALSE    IND   India             NA            NA    
#>  8 DL    Delhi    TRUE     FALSE    IND   India             NA            NA    
#>  9 DN    Dadra a… TRUE     FALSE    IND   India             NA            NA    
#> 10 GA    Goa      TRUE     FALSE    IND   India             NA            NA    
#> # ℹ 24 more rows
#> # ℹ 5 more variables: production_up_2030 <lgl>, production_up_2050 <lgl>,
#> #   cost_USD_t_2020 <dbl>, cost_USD_t_2030 <lgl>, cost_USD_t_2050 <lgl>
  get_ideea_data("coal", nreg = 7, "reserve")
#> # A tibble: 5 × 2
#>   reg7      total_reserve_Mt
#>   <chr>                <dbl>
#> 1 EAST               213174.
#> 2 NORTH                1062.
#> 3 NORTHEAST            1670.
#> 4 SOUTH               27176.
#> 5 WEST               118329.
  get_ideea_data("oil", nreg = 34, "reserve", islands = T)
#> # A tibble: 34 × 4
#>    reg34 mainland oil_reserve_Mt_2021 oil_reserve_GWh_2021
#>    <chr> <lgl>                  <dbl>                <dbl>
#>  1 AN    FALSE                   0                      0 
#>  2 APY   TRUE                    7.33               85248.
#>  3 AR    TRUE                    3.64               42333.
#>  4 AS    TRUE                  153.               1779972.
#>  5 BR    TRUE                    0                      0 
#>  6 CH    TRUE                    0                      0 
#>  7 CT    TRUE                    0                      0 
#>  8 DL    TRUE                    0                      0 
#>  9 DN    TRUE                    0                      0 
#> 10 GA    TRUE                    0                      0 
#> # ℹ 24 more rows
  get_ideea_data("coal", nreg = 7, "cost", agg_fun = mean)
#> # A tibble: 5 × 4
#>   reg7      cost_USD_t_2020 cost_USD_t_2030 cost_USD_t_2050
#>   <chr>               <dbl>           <dbl>           <dbl>
#> 1 EAST                   50             NaN             NaN
#> 2 NORTH                  50             NaN             NaN
#> 3 NORTHEAST              50             NaN             NaN
#> 4 SOUTH                  50             NaN             NaN
#> 5 WEST                   50             NaN             NaN
  get_ideea_data("merra_raw_2014", raw = T) %>% head()
#>                   UTC  locid W10M W50M SWGDN ALBEDO
#> 1 2014-01-01 00:30:00 109878  5.6  6.0    63   0.12
#> 2 2014-01-01 00:30:00 109879  4.4  4.7    71   0.12
#> 3 2014-01-01 00:30:00 110453  6.7  7.3    53   0.13
#> 4 2014-01-01 00:30:00 110454  5.6  6.0    61   0.13
#> 5 2014-01-01 00:30:00 110455  4.2  4.5    51   0.12
#> 6 2014-01-01 00:30:00 111028  8.9  9.9    19   0.10